Solve the 3rd flexbox exercise
I arranged the items in the navbar. Used a little
different approach than the solution as I pushed
one of the parts of the header to right using margin
auto. I also resized the button which is not in the
solutions.

However my solution looks the same as the
desired one, so I'm just gonna take it as one of those
examples in coding of the fact that there is several
ways to achieve the same solution.