Finish the 4th flexbox exercise
I aligned the items according to the desired look.
I used some slightly different ways than the solution.
For example I used the top margin of the lower
container to set it rather than the bottom of the
header but that is OK I assume.

I also had to group some stuff into containers in
the index.html to use flexbox.